requirements.txt
langchain
langchain-community
ipykernel
python-dotenv
pymupdf
pypdf
beautifulsoup4
wikipedia
arxiv
langchain-text-splitters
langchain-google-genai
google-genai
google-generativeai
chromadb
huggingface-hub
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma


############
ipywidgets
tqdm
langchain
langchain-core
langchain-community
langchain-google-genai
python-dotenv
streamlit
streamlit-chat


